Updated FOSDEM with talks by Red Hat employees
This is based on the Pentabarf export from FOSDEM.org as of 2016-01-11, filtered against the current LDAP information as of this date.

As it matches against names, there may be some false positives or negatives — it needs to be vetted and updated if/where errors exist. Based on experience from previous years, it's probably pretty accurate already, however. (Also, corrections of people from previous years — mainly diacritics and other non-ASCII characters, as well as nicknames — are already included, so it should help to minimize issues.)